requirements.txt
# RAG Chatbot Dependencies
# Core Framework
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.0
langchain-anthropic==0.3.0  # For Claude Sonnet 4.5

# Vector Store
chromadb>=0.6.0

# Embeddings
openai==1.58.1
sentence-transformers==3.3.1

# Document Loaders
pymupdf==1.25.1  # PyMuPDF for superior PDF parsing
python-docx==1.1.2
# unstructured not needed - using PyMuPDF directly

# Environment & Configuration
python-dotenv==1.0.1

# UI (Optional)
streamlit==1.41.1

# Evaluation & Testing
pytest==8.3.4
pandas==2.2.3
numpy==2.2.1

# Utilities
tiktoken==0.8.0
tqdm==4.67.1
anthropic==0.42.0  # Anthropic Python SDK
